+++ js/input/Mouse.js
@@ -9,6 +9,7 @@
 define( function( require ) {
   'use strict';
   
+  var inherit = require( 'PHET_CORE/inherit' );
   var scenery = require( 'SCENERY/scenery' );
   
   var Pointer = require( 'SCENERY/input/Pointer' ); // inherits from Pointer
@@ -30,9 +31,7 @@ define( function( require ) {
   };
   var Mouse = scenery.Mouse;
   
-  Mouse.prototype = _.extend( {}, Pointer.prototype, {
-    constructor: Mouse,
-    
+  inherit( Mouse, Pointer, {
     down: function( point, event ) {
       this.point = point;
       switch( event.button ) {