+++ js/input/Pen.js
@@ -9,6 +9,7 @@
 define( function( require ) {
   'use strict';
   
+  var inherit = require( 'PHET_CORE/inherit' );
   var scenery = require( 'SCENERY/scenery' );
   
   var Pointer = require( 'SCENERY/input/Pointer' ); // extends Pointer
@@ -25,9 +26,7 @@ define( function( require ) {
   };
   var Pen = scenery.Pen;
   
-  Pen.prototype = _.extend( {}, Pointer.prototype, {
-    constructor: Pen,
-    
+  inherit( Pen, Pointer, {
     move: function( point, event ) {
       this.point = point;
     },