+++ js/input/Touch.js
@@ -11,6 +11,7 @@
 define( function( require ) {
   'use strict';
   
+  var inherit = require( 'PHET_CORE/inherit' );
   var scenery = require( 'SCENERY/scenery' );
   
   var Pointer = require( 'SCENERY/input/Pointer' ); // extends Pointer
@@ -27,9 +28,7 @@ define( function( require ) {
   };
   var Touch = scenery.Touch;
   
-  Touch.prototype = _.extend( {}, Pointer.prototype, {
-    constructor: Touch,
-    
+  inherit( Touch, Pointer, {
     move: function( point, event ) {
       this.point = point;
     },