+++ js/main.js
@@ -47,6 +47,7 @@ define( [
     'SCENERY/nodes/Text',
     'SCENERY/nodes/VBox',
     
+    'SCENERY/util/AccessibilityPeer',
     'SCENERY/util/BasicNodeEvents',
     'SCENERY/util/CanvasContextWrapper',
     'SCENERY/util/Color',